"Feature Complete" Prototype (Pre-Alpha still because no menus or boss?)
Added scripts and assets for meter, established basic gameplay loop pre-boss, tweaked numbers to make a usable "demo" for team.

"Demo" still isn't a full level with boss or menus, and there are absolutely no bells or whistles to speak of, but this is the first big step in the direction of a completed game. Yay!